refactor: chat widget 로직 위치 변경
